Extract search operations into search module
